Disable deno 'no-var' rule
I agree with Kyle Simpson (You Don't Know JS guy) that we should know
*and understand* the features of a language and use them correctly.
`var` is not “worse” then `let` and `const` as functions created with
the `function` keyword (declarations or expressions) are not inferior to
arrow functions. They behave differently for a number of reasons and
they all have their pros and cons.

For example, some algorithms benefit from a variable to *not be*
block-scoped, but instead, function-scoped.